aktualizace cv05
oprava cv05
