+++ www/smap01/cv05/Database/productsDB.db
@@ -1,0 +1,1 @@
+Table;2000 Kč
