+++ www/smap01/cv05/Database/usersDB.db
@@ -1,0 +1,1 @@
+Patrik Šmátrala;smap01@vse.cz
