+++ www/smap01/cv05/database.php
@@ -4,21 +4,23 @@
 
 abstract class Database implements databaseOperations
 {
-    protected $dbPath= './database/';
+    protected $dbPath = './database/';
     protected $dbExtension = '.db';
     protected $separator = ';';
 
     public function __construct()
     {
-        $database=fopen($this->dbPath.static::class.$this->dbExtension,'w');
+        $database = fopen($this->dbPath . static::class . $this->dbExtension, 'w');
         fclose($database);
-        echo '<br>Database class '. static::class. ' was initialized.</br>'. PHP_EOL;
+        echo '<br>Database class ' . static::class . ' was initialized.</br>' . PHP_EOL;
     }
-    public function __toString() {
-        return "Database ". static::class." with a path '<i>".$this->dbPath.$this->dbExtension."</i>' with a delimeter '<i>".$this->separator."</i>'.".PHP_EOL;
+    public function __toString()
+    {
+        return "Database " . static::class . " with a path '<i>" . $this->dbPath . $this->dbExtension . "</i>' with a delimeter '<i>" . $this->separator . "</i>'." . PHP_EOL;
     }
-    public function getFilePath(){
-        return $this->dbPath.static::class.$this->dbExtension;
+    public function getFilePath()
+    {
+        return $this->dbPath . static::class . $this->dbExtension;
     }
 }
